convert LimitSurface, FTImporter, and BasicStatistics
