+++ framework/Models/Factory.py
@@ -22,43 +22,43 @@ from .ROM           import ROM
 from .ExternalModel import ExternalModel
 from .Code          import Code
 from .EnsembleModel import EnsembleModel
-from .PostProcessor import PostProcessor
 from .HybridModels  import HybridModel
 from .HybridModels  import LogicalModel
 
 #### PostProcessors
-from .PostProcessorBase import PostProcessorBase
-from .Metric import Metric
-from .ETImporter import ETImporter
-from .DataMining import DataMining
-from .SafestPoint import SafestPoint
-from .LimitSurface import LimitSurface
-from .ValueDuration import ValueDuration
-from .SampleSelector import SampleSelector
-from .ImportanceRank import ImportanceRank
-from .CrossValidation import CrossValidation
-from .BasicStatistics import BasicStatistics
-from .LimitSurfaceIntegral import LimitSurfaceIntegral
-from .FastFourierTransform import FastFourierTransform
-from .ExternalPostProcessor import ExternalPostProcessor
-from .InterfacedPostProcessor import InterfacedPostProcessor
-from .TopologicalDecomposition import TopologicalDecomposition
-from .FTImporter import FTImporter
-from .DataClassifier import DataClassifier
-from .ComparisonStatisticsModule import ComparisonStatistics
-from .RealizationAverager import RealizationAverager
-from .ParetoFrontierPostProcessor import ParetoFrontier
-from .MCSimporter import MCSImporter
-from .EconomicRatio import EconomicRatio
-# from .RavenOutput import RavenOutput # deprecated for now
+from .PostProcessors import PostProcessor
+from .PostProcessors import FTImporter
 
-## These utilize the optional prequisite library PySide, so don't error if they
-## do not import appropriately.
-try:
-  from .TopologicalDecomposition import QTopologicalDecomposition
-  from .DataMining import QDataMining
-except ImportError:
-  pass
+# from .PostProcessors import Metric
+# from .PostProcessors import ETImporter
+# from .PostProcessors import DataMining
+# from .PostProcessors import SafestPoint
+# from .PostProcessors import LimitSurface
+# from .PostProcessors import ValueDuration
+# from .PostProcessors import SampleSelector
+# from .PostProcessors import ImportanceRank
+# from .PostProcessors import CrossValidation
+# from .PostProcessors import BasicStatistics
+# from .PostProcessors import LimitSurfaceIntegral
+# from .PostProcessors import FastFourierTransform
+# from .PostProcessors import ExternalPostProcessor
+# from .PostProcessors import InterfacedPostProcessor
+# from .PostProcessors import TopologicalDecomposition
+# from .PostProcessors import DataClassifier
+# from .PostProcessors.ComparisonStatisticsModule import ComparisonStatistics
+# from .PostProcessors import RealizationAverager
+# from .PostProcessors.ParetoFrontierPostProcessor import ParetoFrontier
+# from .PostProcessors.MCSimporter import MCSImporter
+# from .PostProcessors import EconomicRatio
+# # from .PostProcessors import RavenOutput # deprecated for now
+#
+# ## These utilize the optional prequisite library PySide, so don't error if they
+# ## do not import appropriately.
+# try:
+#   from .PostProcessors.TopologicalDecomposition import QTopologicalDecomposition
+#   from .PostProcessors.DataMining import QDataMining
+# except ImportError:
+#   pass
 
 __base = 'Model'
 __interFaceDict = {}
@@ -67,14 +67,14 @@ for classObj in utils.getAllSubclasses(eval(__base)):
   key = classObj.__name__
   __interFaceDict[key] = classObj
 
-## Adding aliases for certain classes that are exposed to the user.
-__interFaceDict['External'] = ExternalPostProcessor
-try:
-  __interFaceDict['TopologicalDecomposition' ] = QTopologicalDecomposition
-  __interFaceDict['DataMining'               ] = QDataMining
-except NameError:
-  ## The correct names should already be used for these classes otherwise
-  pass
+# ## Adding aliases for certain classes that are exposed to the user.
+# __interFaceDict['External'] = ExternalPostProcessor
+# try:
+#   __interFaceDict['TopologicalDecomposition' ] = QTopologicalDecomposition
+#   __interFaceDict['DataMining'               ] = QDataMining
+# except NameError:
+#   ## The correct names should already be used for these classes otherwise
+#   pass
 
 #here the class methods are called to fill the information about the usage of the classes
 for classType in __interFaceDict.values():
