+++ framework/Models/__init__.py
@@ -27,47 +27,47 @@ from .ROM import ROM
 from .ExternalModel import ExternalModel
 from .Code import Code
 from .EnsembleModel import EnsembleModel
-from .PostProcessor import PostProcessor
 from .HybridModels import HybridModel
 from .HybridModels import LogicalModel
 
 #### PostProcessors
-from .PostProcessorBase import PostProcessorBase
-from .Metric import Metric
-from .ETImporter import ETImporter
-from .DataMining import DataMining
-from .SafestPoint import SafestPoint
-from .LimitSurface import LimitSurface
-from .ValueDuration import ValueDuration
-from .SampleSelector import SampleSelector
-from .ImportanceRank import ImportanceRank
-from .CrossValidation import CrossValidation
-from .BasicStatistics import BasicStatistics
-from .LimitSurfaceIntegral import LimitSurfaceIntegral
-from .FastFourierTransform import FastFourierTransform
-from .ExternalPostProcessor import ExternalPostProcessor
-from .InterfacedPostProcessor import InterfacedPostProcessor
-from .TopologicalDecomposition import TopologicalDecomposition
-from .FTImporter import FTImporter
-from .DataClassifier import DataClassifier
-from .ComparisonStatisticsModule import ComparisonStatistics
-from .RealizationAverager import RealizationAverager
-from .ParetoFrontierPostProcessor import ParetoFrontier
-from .MCSimporter import MCSImporter
-from .EconomicRatio import EconomicRatio
-# from .RavenOutput import RavenOutput # deprecated for now
+from .PostProcessors import PostProcessor
+from .PostProcessors import FTImporter
 
-additionalModules = []
-## These utilize the optional prequisite library PySide, so don't error if they
-## do not import appropriately.
-try:
-  from .TopologicalDecomposition import QTopologicalDecomposition
-  from .DataMining import QDataMining
-  additionalModules.append(QTopologicalDecomposition)
-  additionalModules.append(QDataMining)
-except ImportError:
-  ## User most likely does not have PySide installed and working
-  pass
+# from .PostProcessors import Metric
+# from .PostProcessors import ETImporter
+# from .PostProcessors import DataMining
+# from .PostProcessors import SafestPoint
+# from .PostProcessors import LimitSurface
+# from .PostProcessors import ValueDuration
+# from .PostProcessors import SampleSelector
+# from .PostProcessors import ImportanceRank
+# from .PostProcessors import CrossValidation
+# from .PostProcessors import BasicStatistics
+# from .PostProcessors import LimitSurfaceIntegral
+# from .PostProcessors import FastFourierTransform
+# from .PostProcessors import ExternalPostProcessor
+# from .PostProcessors import InterfacedPostProcessor
+# from .PostProcessors import TopologicalDecomposition
+# from .PostProcessors import DataClassifier
+# from .PostProcessors.ComparisonStatisticsModule import ComparisonStatistics
+# from .PostProcessors import RealizationAverager
+# from .PostProcessors.ParetoFrontierPostProcessor import ParetoFrontier
+# from .PostProcessors.MCSimporter import MCSImporter
+# from .PostProcessors import EconomicRatio
+# # from .PostProcessors import RavenOutput # deprecated for now
+#
+# additionalModules = []
+# ## These utilize the optional prequisite library PySide, so don't error if they
+# ## do not import appropriately.
+# try:
+#   from .PostProcessors.TopologicalDecomposition import QTopologicalDecomposition
+#   from .PostProcessors.DataMining import QDataMining
+#   additionalModules.append(QTopologicalDecomposition)
+#   additionalModules.append(QDataMining)
+# except ImportError:
+#   ## User most likely does not have PySide installed and working
+#   pass
 
 
 ## [ Add new class here ]
@@ -77,34 +77,34 @@ from .Factory import returnInstance
 from .Factory import needsRunInfo
 from .Factory import validate
 
-# We should not really need this as we do not use wildcard imports
-__all__ = ['Model',
-           'Dummy',
-           'ROM',
-           'ExternalModel',
-           'Code',
-           'EnsembleModel',
-           'PostProcessor',
-           'HybridModel',
-           'LogicalModel',
-           'BasicStatistics',
-           'ComparisonStatistics',
-           'ExternalPostProcessor',
-           'ImportanceRank',
-           'InterfacedPostProcessor',
-           'LimitSurface',
-           'LimitSurfaceIntegral',
-           'SafestPoint',
-           'TopologicalDecomposition',
-           'DataMining',
-           'Metric',
-           'CrossValidation',
-           'ValueDuration',
-           'FastFourierTransform',
-           'FTImporter',
-           'DataClassifier',
-           'SampleSelector',
-           'ETImporter',
-           'RealizationAverager',
-           'ParetoFrontier']+ additionalModules
-#           'RavenOutput', # deprecated for now
+# # We should not really need this as we do not use wildcard imports
+# __all__ = ['Model',
+#            'Dummy',
+#            'ROM',
+#            'ExternalModel',
+#            'Code',
+#            'EnsembleModel',
+#            'PostProcessor',
+#            'HybridModel',
+#            'LogicalModel',
+#            'BasicStatistics',
+#            'ComparisonStatistics',
+#            'ExternalPostProcessor',
+#            'ImportanceRank',
+#            'InterfacedPostProcessor',
+#            'LimitSurface',
+#            'LimitSurfaceIntegral',
+#            'SafestPoint',
+#            'TopologicalDecomposition',
+#            'DataMining',
+#            'Metric',
+#            'CrossValidation',
+#            'ValueDuration',
+#            'FastFourierTransform',
+#            'FTImporter',
+#            'DataClassifier',
+#            'SampleSelector',
+#            'ETImporter',
+#            'RealizationAverager',
+#            'ParetoFrontier']+ additionalModules
+# #           'RavenOutput', # deprecated for now
